+++ pyccel/codegen/printing/pycode.py
@@ -510,6 +510,14 @@ class PythonCodePrinter(CodePrinter):
         else:
             raise NotImplementedError("The shape access function seems to be poorly defined.")
 
+    def _print_PythonRound(self, expr):
+        arg = self._print(expr.arg)
+        if expr.ndigits:
+            ndigits = self._print(expr.ndigits)
+            return f'round({arg}, {ndigits})'
+        else:
+            return f'round({arg})'
+
     def _print_PyccelArraySize(self, expr):
         arg = self._print(expr.arg)
         name = self._get_numpy_name(expr)
@@ -837,7 +845,11 @@ class PythonCodePrinter(CodePrinter):
         else:
             method_args = ', '.join(self._print(a) for a in expr.args)
 
-        return f"{list_obj}.{method_name}({method_args})\n"
+        code = f"{list_obj}.{method_name}({method_args})"
+        if isinstance(expr.class_type, VoidType):
+            return code + '\n'
+        else:
+            return code
 
     def _print_DictMethod(self, expr):
         method_name = expr.name